Refactor vm
