+++ vm/compiler/src/ledger/vm/authorize.rs
@@ -1,0 +1,52 @@
+// Copyright (C) 2019-2022 Aleo Systems Inc.
+// This file is part of the snarkVM library.
+
+// The snarkVM library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The snarkVM library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
+
+use super::*;
+
+impl<N: Network> VM<N> {
+    /// Authorizes a call to the program function for the given inputs.
+    #[inline]
+    pub fn authorize<R: Rng + CryptoRng>(
+        &self,
+        private_key: &PrivateKey<N>,
+        program_id: &ProgramID<N>,
+        function_name: Identifier<N>,
+        inputs: &[Value<N>],
+        rng: &mut R,
+    ) -> Result<Authorization<N>> {
+        // Compute the core logic.
+        macro_rules! logic {
+            ($process:expr, $network:path, $aleo:path) => {{
+                let inputs = inputs.to_vec();
+
+                // Prepare the inputs.
+                let private_key = cast_ref!(&private_key as PrivateKey<$network>);
+                let program_id = cast_ref!(&program_id as ProgramID<$network>);
+                let function_name = cast_ref!(function_name as Identifier<$network>);
+                let inputs = cast_ref!(inputs as Vec<Value<$network>>);
+
+                // Compute the authorization.
+                let authorization =
+                    $process.authorize::<$aleo, _>(private_key, program_id, function_name.clone(), inputs, rng)?;
+
+                // Return the authorization.
+                Ok(cast_ref!(authorization as Authorization<N>).clone())
+            }};
+        }
+        // Process the logic.
+        process!(self, logic)
+    }
+}
