+++ vm/compiler/src/ledger/vm/mod.rs
@@ -16,6 +16,12 @@
 
 mod helpers;
 
+mod authorize;
+mod deploy;
+mod execute;
+mod finalize;
+mod verify;
+
 use crate::{
     cast_ref,
     ledger::{
@@ -118,315 +124,6 @@ impl<N: Network> VM<N> {
                 false
             }
         }
-    }
-
-    /// Deploys a program with the given program ID.
-    #[inline]
-    pub fn deploy<R: Rng + CryptoRng>(&self, program: &Program<N>, rng: &mut R) -> Result<Deployment<N>> {
-        // Compute the core logic.
-        macro_rules! logic {
-            ($process:expr, $network:path, $aleo:path) => {{
-                // Prepare the program.
-                let program = cast_ref!(&program as Program<$network>);
-
-                // Compute the deployment.
-                let deployment = $process.deploy::<$aleo, _>(program, rng)?;
-
-                // Prepare the return.
-                let deployment = cast_ref!(deployment as Deployment<N>).clone();
-                // Return the deployment.
-                Ok(deployment)
-            }};
-        }
-        // Process the logic.
-        process!(self, logic)
-    }
-
-    /// Authorizes a call to the program function for the given inputs.
-    #[inline]
-    pub fn authorize<R: Rng + CryptoRng>(
-        &self,
-        private_key: &PrivateKey<N>,
-        program_id: &ProgramID<N>,
-        function_name: Identifier<N>,
-        inputs: &[Value<N>],
-        rng: &mut R,
-    ) -> Result<Authorization<N>> {
-        // Compute the core logic.
-        macro_rules! logic {
-            ($process:expr, $network:path, $aleo:path) => {{
-                let inputs = inputs.to_vec();
-
-                // Prepare the inputs.
-                let private_key = cast_ref!(&private_key as PrivateKey<$network>);
-                let program_id = cast_ref!(&program_id as ProgramID<$network>);
-                let function_name = cast_ref!(function_name as Identifier<$network>);
-                let inputs = cast_ref!(inputs as Vec<Value<$network>>);
-
-                // Compute the authorization.
-                let authorization =
-                    $process.authorize::<$aleo, _>(private_key, program_id, function_name.clone(), inputs, rng)?;
-
-                // Return the authorization.
-                Ok(cast_ref!(authorization as Authorization<N>).clone())
-            }};
-        }
-        // Process the logic.
-        process!(self, logic)
-    }
-
-    /// Executes a call to the program function for the given inputs.
-    #[inline]
-    pub fn execute<R: Rng + CryptoRng>(
-        &self,
-        authorization: Authorization<N>,
-        rng: &mut R,
-    ) -> Result<(Response<N>, Execution<N>)> {
-        // Compute the core logic.
-        macro_rules! logic {
-            ($process:expr, $network:path, $aleo:path) => {{
-                // Prepare the authorization.
-                let authorization = cast_ref!(authorization as Authorization<$network>);
-
-                // Execute the call.
-                let (response, execution) = $process.execute::<$aleo, _>(authorization.clone(), rng)?;
-
-                // Prepare the return.
-                let response = cast_ref!(response as Response<N>).clone();
-                let execution = cast_ref!(execution as Execution<N>).clone();
-                // Return the response and execution.
-                Ok((response, execution))
-            }};
-        }
-        // Process the logic.
-        process!(self, logic)
-    }
-
-    /// Returns an additional fee for the given private key, credits record, and additional fee amount (in gates).
-    #[inline]
-    pub fn execute_additional_fee<R: Rng + CryptoRng>(
-        &self,
-        private_key: &PrivateKey<N>,
-        credits: Record<N, Plaintext<N>>,
-        additional_fee_in_gates: u64,
-        rng: &mut R,
-    ) -> Result<(Response<N>, AdditionalFee<N>)> {
-        // Compute the core logic.
-        macro_rules! logic {
-            ($process:expr, $network:path, $aleo:path) => {{
-                type RecordPlaintext<NetworkMacro> = Record<NetworkMacro, Plaintext<NetworkMacro>>;
-
-                // Prepare the private key and credits record.
-                let private_key = cast_ref!(&private_key as PrivateKey<$network>);
-                let credits = cast_ref!(credits as RecordPlaintext<$network>);
-
-                // Execute the call to additional fee.
-                let (response, additional_fee) = $process.execute_additional_fee::<$aleo, _>(
-                    private_key,
-                    credits.clone(),
-                    additional_fee_in_gates,
-                    rng,
-                )?;
-
-                // Prepare the return.
-                let response = cast_ref!(response as Response<N>).clone();
-                let additional_fee = cast_ref!(additional_fee as AdditionalFee<N>).clone();
-                // Return the response and additional fee.
-                Ok((response, additional_fee))
-            }};
-        }
-        // Process the logic.
-        process!(self, logic)
-    }
-
-    /// Verifies the transaction in the VM.
-    #[inline]
-    pub fn verify(&self, transaction: &Transaction<N>) -> bool {
-        // Compute the Merkle root of the transaction.
-        match transaction.to_root() {
-            // Ensure the transaction ID is correct.
-            Ok(root) => {
-                if *transaction.id() != root {
-                    warn!("Incorrect transaction ID ({})", transaction.id());
-                    return false;
-                }
-            }
-            Err(error) => {
-                warn!("Failed to compute the Merkle root of the transaction: {error}\n{transaction}");
-                return false;
-            }
-        };
-
-        // Ensure there are no duplicate transition IDs.
-        if has_duplicates(transaction.transition_ids()) {
-            warn!("Found duplicate transition in the transactions list");
-            return false;
-        }
-
-        // Ensure there are no duplicate transition public keys.
-        if has_duplicates(transaction.transition_public_keys()) {
-            warn!("Found duplicate transition public keys in the transactions list");
-            return false;
-        }
-
-        // Ensure there are no duplicate serial numbers.
-        if has_duplicates(transaction.serial_numbers()) {
-            warn!("Found duplicate serial numbers in the transactions list");
-            return false;
-        }
-
-        // Ensure there are no duplicate commitments.
-        if has_duplicates(transaction.commitments()) {
-            warn!("Found duplicate commitments in the transactions list");
-            return false;
-        }
-
-        // Ensure there are no duplicate nonces.
-        if has_duplicates(transaction.nonces()) {
-            warn!("Found duplicate nonces in the transactions list");
-            return false;
-        }
-
-        match transaction {
-            Transaction::Deploy(_, deployment, additional_fee) => {
-                // Check the deployment size.
-                if let Err(error) = Transaction::check_deployment_size(deployment) {
-                    warn!("Invalid transaction size (deployment): {error}");
-                    return false;
-                }
-                // Verify the deployment.
-                self.verify_deployment(deployment)
-                    // Verify the additional fee.
-                    && self.verify_additional_fee(additional_fee)
-            }
-            Transaction::Execute(_, execution, additional_fee) => {
-                // Check the deployment size.
-                if let Err(error) = Transaction::check_execution_size(execution) {
-                    warn!("Invalid transaction size (execution): {error}");
-                    return false;
-                }
-
-                // Verify the additional fee, if it exists.
-                let check_additional_fee = match additional_fee {
-                    Some(additional_fee) => self.verify_additional_fee(additional_fee),
-                    None => true,
-                };
-
-                // Verify the execution.
-                self.verify_execution(execution)
-                    // Verify the additional fee.
-                    && check_additional_fee
-            }
-        }
-    }
-
-    /// Verifies the given deployment.
-    #[inline]
-    fn verify_deployment(&self, deployment: &Deployment<N>) -> bool {
-        // Compute the core logic.
-        macro_rules! logic {
-            ($process:expr, $network:path, $aleo:path) => {{
-                let task = || {
-                    // Prepare the deployment.
-                    let deployment = cast_ref!(&deployment as Deployment<$network>);
-                    // Initialize an RNG.
-                    let rng = &mut rand::thread_rng();
-                    // Verify the deployment.
-                    $process.verify_deployment::<$aleo, _>(&deployment, rng)
-                };
-                task()
-            }};
-        }
-
-        // Process the logic.
-        match process!(self, logic) {
-            Ok(()) => true,
-            Err(error) => {
-                warn!("Deployment verification failed: {error}");
-                false
-            }
-        }
-    }
-
-    /// Verifies the given execution.
-    #[inline]
-    fn verify_execution(&self, execution: &Execution<N>) -> bool {
-        // Compute the core logic.
-        macro_rules! logic {
-            ($process:expr, $network:path, $aleo:path) => {{
-                let task = || {
-                    // Prepare the execution.
-                    let execution = cast_ref!(&execution as Execution<$network>);
-                    // Verify the execution.
-                    $process.verify_execution(execution)
-                };
-                task()
-            }};
-        }
-
-        // Process the logic.
-        match process!(self, logic) {
-            Ok(()) => true,
-            Err(error) => {
-                warn!("Execution verification failed: {error}");
-                false
-            }
-        }
-    }
-
-    /// Verifies the given additional fee.
-    #[inline]
-    fn verify_additional_fee(&self, additional_fee: &AdditionalFee<N>) -> bool {
-        // Compute the core logic.
-        macro_rules! logic {
-            ($process:expr, $network:path, $aleo:path) => {{
-                let task = || {
-                    // Prepare the additional fee.
-                    let additional_fee = cast_ref!(&additional_fee as AdditionalFee<$network>);
-                    // Verify the additional fee.
-                    $process.verify_additional_fee(additional_fee)
-                };
-                task()
-            }};
-        }
-
-        // Process the logic.
-        match process!(self, logic) {
-            Ok(()) => true,
-            Err(error) => {
-                warn!("Additional fee verification failed: {error}");
-                false
-            }
-        }
-    }
-
-    /// Finalizes the transaction into the VM.
-    /// This method assumes the given transaction **is valid**.
-    #[inline]
-    pub fn finalize(&mut self, transaction: &Transaction<N>) -> Result<()> {
-        // Ensure the transaction is valid.
-        ensure!(self.verify(transaction), "Invalid transaction: failed to verify");
-        // Finalize the transaction.
-        match transaction {
-            Transaction::Deploy(_, deployment, _) => self.finalize_deployment(deployment),
-            Transaction::Execute(_, _execution, _) => Ok(()), // self.finalize_execution(execution),
-        }
-    }
-
-    /// Adds the newly-deployed program into the VM.
-    #[inline]
-    fn finalize_deployment(&mut self, deployment: &Deployment<N>) -> Result<()> {
-        // Compute the core logic.
-        macro_rules! logic {
-            ($process:expr, $network:path, $aleo:path) => {{
-                // Prepare the deployment.
-                let deployment = cast_ref!(&deployment as Deployment<$network>);
-                // Finalize the deployment.
-                $process.finalize_deployment(deployment)
-            }};
-        }
-        // Process the logic.
-        process_mut!(self, logic)
     }
 }
 
@@ -567,79 +264,5 @@ function compute:
                 transaction
             })
             .clone()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{ledger::vm::test_helpers::sample_program, VM};
-    use console::network::Testnet3;
-    use snarkvm_utilities::test_crypto_rng;
-
-    type CurrentNetwork = Testnet3;
-
-    #[test]
-    fn test_verify() {
-        let vm = VM::<CurrentNetwork>::new().unwrap();
-
-        // Fetch a deployment transaction.
-        let deployment_transaction = crate::ledger::vm::test_helpers::sample_deployment_transaction();
-        // Ensure the transaction verifies.
-        assert!(vm.verify(&deployment_transaction));
-
-        // Fetch a execution transaction.
-        let execution_transaction = crate::ledger::vm::test_helpers::sample_execution_transaction();
-        // Ensure the transaction verifies.
-        assert!(vm.verify(&execution_transaction));
-    }
-
-    #[test]
-    fn test_verify_deployment() {
-        let rng = &mut test_crypto_rng();
-        let vm = VM::<CurrentNetwork>::new().unwrap();
-
-        // Fetch the program from the deployment.
-        let program = sample_program();
-
-        // Deploy the program.
-        let deployment = vm.deploy(&program, rng).unwrap();
-
-        // Ensure the deployment is valid.
-        assert!(vm.verify_deployment(&deployment));
-    }
-
-    #[test]
-    fn test_finalize() {
-        let mut vm = VM::<CurrentNetwork>::new().unwrap();
-
-        // Fetch a deployment transaction.
-        let deployment_transaction = crate::ledger::vm::test_helpers::sample_deployment_transaction();
-
-        // Finalize the transaction.
-        vm.finalize(&deployment_transaction).unwrap();
-
-        // Ensure the VM can't redeploy the same transaction.
-        assert!(vm.finalize(&deployment_transaction).is_err());
-    }
-
-    #[test]
-    fn test_finalize_deployment() {
-        let rng = &mut test_crypto_rng();
-        let mut vm = VM::<CurrentNetwork>::new().unwrap();
-
-        // Fetch the program from the deployment.
-        let program = sample_program();
-
-        // Deploy the program.
-        let deployment = vm.deploy(&program, rng).unwrap();
-
-        // Ensure the program does not exists.
-        assert!(!vm.contains_program(program.id()));
-
-        // Finalize the deployment.
-        vm.finalize_deployment(&deployment).unwrap();
-
-        // Ensure the program exists.
-        assert!(vm.contains_program(program.id()));
     }
 }
